# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pavilion community website API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /special-contributions:
    get:
      operationId: listSpecialContributions
      tags: [special-contributions]
      summary: List special contributions (admin only)
      responses:
        "200":
          description: Special contributions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpecialContribution"
        "401":
          description: Not signed in
        "403":
          description: Not an admin
    post:
      operationId: createSpecialContribution
      tags: [special-contributions]
      summary: Announce a special contribution (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecialContributionInput"
      responses:
        "201":
          description: Special contribution created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecialContribution"
        "401":
          description: Not signed in
        "403":
          description: Not an admin
  /special-contributions/{id}:
    delete:
      operationId: deleteSpecialContribution
      tags: [special-contributions]
      summary: Delete a special contribution (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Special contribution deleted
        "401":
          description: Not signed in
        "403":
          description: Not an admin
        "404":
          description: Special contribution not found
  /vendor-bills:
    get:
      operationId: listVendorBills
      tags: [vendor-bills]
      summary: List vendor bills (maintenance expenses), with paid-so-far and status (admin only)
      responses:
        "200":
          description: Vendor bills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VendorBill"
        "401":
          description: Not signed in
        "403":
          description: Not an admin
    post:
      operationId: createVendorBill
      tags: [vendor-bills]
      summary: Add a vendor bill (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorBillInput"
      responses:
        "201":
          description: Vendor bill created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorBill"
        "401":
          description: Not signed in
        "403":
          description: Not an admin
        "404":
          description: Unknown vendor or expense category
  /vendor-bills/{id}:
    delete:
      operationId: deleteVendorBill
      tags: [vendor-bills]
      summary: Delete a vendor bill (admin only) — fails if payments have been recorded against it
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Vendor bill deleted
        "401":
          description: Not signed in
        "403":
          description: Not an admin
        "404":
          description: Vendor bill not found
        "409":
          description: Payments have already been recorded against this bill
...
components:
  schemas:
...
    SpecialContribution:
      type: object
      required: [id, title, description, amountPaise, dueDate, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        amountPaise:
          type: integer
        dueDate:
          type: string
        createdAt:
          type: string
          format: date-time
    SpecialContributionInput:
      type: object
      required: [title, amountPaise, dueDate]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        amountPaise:
          type: integer
          minimum: 0
        dueDate:
          type: string
          minLength: 1
    VendorBill:
      type: object
      required:
        [
          id,
          vendorId,
          vendorName,
          expenseCategoryId,
          expenseCategoryName,
          billNumber,
          billDate,
          amountPaise,
          description,
          paidAmountPaise,
          status,
          createdAt,
        ]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        vendorName:
          type: string
        expenseCategoryId:
          type: integer
        expenseCategoryName:
          type: string
        billNumber:
          type: string
        billDate:
          type: string
        amountPaise:
          type: integer
        description:
          type: ["string", "null"]
        paidAmountPaise:
          type: integer
        status:
          type: string
          enum: [unpaid, partially_paid, paid]
        createdAt:
          type: string
          format: date-time
    VendorBillInput:
      type: object
      required: [vendorId, expenseCategoryId, billNumber, billDate, amountPaise]
      properties:
        vendorId:
          type: integer
        expenseCategoryId:
          type: integer
        billNumber:
          type: string
          minLength: 1
        billDate:
          type: string
          minLength: 1
        amountPaise:
          type: integer
          minimum: 1
        description:
          type: string
...