# lib/api-zod/src/generated/types/specialContributionInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Pavilion community website API
 * OpenAPI spec version: 0.1.0
 */

export interface SpecialContributionInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  /** @minimum 0 */
  amountPaise: number;
  /** @minLength 1 */
  dueDate: string;
}