# lib/api-zod/src/generated/types/vendorBillInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Pavilion community website API
 * OpenAPI spec version: 0.1.0
 */

export interface VendorBillInput {
  vendorId: number;
  expenseCategoryId: number;
  /** @minLength 1 */
  billNumber: string;
  /** @minLength 1 */
  billDate: string;
  /** @minimum 1 */
  amountPaise: number;
  description?: string;
}